# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Sistema de Panificación
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ventas:
    get:
      operationId: listarVentas
      tags: [ventas]
      summary: List all sales
      parameters:
        - name: fechaDesde
          in: query
          required: false
          schema:
            type: string
        - name: fechaHasta
          in: query
          required: false
          schema:
            type: string
        - name: vendedor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venta"
    post:
      operationId: registrarVenta
      tags: [ventas]
      summary: Register a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrarVentaBody"
      responses:
        "201":
          description: Sale registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venta"
  /ventas/{id}:
    get:
      operationId: obtenerVenta
      tags: [ventas]
      summary: Get sale by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venta"
...
components:
  schemas:
...
    Venta:
      type: object
      properties:
        id:
          type: integer
        vendedor:
          type: string
        fecha:
          type: string
        total:
          type: number
        ganancia:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemVenta"
      required: [id, vendedor, fecha, total, ganancia, items]
    ItemVenta:
      type: object
      properties:
        id:
          type: integer
        ventaId:
          type: integer
        productoCodigo:
          type: string
        productoNombre:
          type: string
        cantidad:
          type: integer
        precioUnitario:
          type: number
        precioCosto:
          type: number
        subtotal:
          type: number
      required: [id, ventaId, productoCodigo, productoNombre, cantidad, precioUnitario, precioCosto, subtotal]
    RegistrarVentaBody:
      type: object
      properties:
        vendedor:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemVentaInput"
      required: [vendedor, items]
    ItemVentaInput:
      type: object
      properties:
        productoCodigo:
          type: string
        cantidad:
          type: integer
      required: [productoCodigo, cantidad]
...